# openapi.json
{"openapi":"3.1.0","info":{"title":"Sports Pick Tracker","description":"Track top sports pick influencers and get AI-powered consensus recommendations.","version":"1.0.0"},"paths":{"/health":{"get":{"summary":"Health","operationId":"health_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/influencers":{"get":{"summary":"List Influencers","operationId":"list_influencers_influencers_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}},{"name":"min_picks","in":"query","required":false,"schema":{"type":"integer","minimum":0,"default":0,"title":"Min Picks"}},{"name":"sort_by","in":"query","required":false,"schema":{"type":"string","pattern":"^(elo_score|accuracy_rate|total_picks|follower_count|avg_clv)$","default":"elo_score","title":"Sort By"}},{"name":"platform","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Platform"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/influencers/{influencer_id}":{"get":{"summary":"Get Influencer","operationId":"get_influencer_influencers__influencer_id__get","parameters":[{"name":"influencer_id","in":"path","required":true,"schema":{"type":"string","title":"Influencer Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/matches":{"get":{"summary":"List Matches","operationId":"list_matches_matches_get","parameters":[{"name":"stage","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Stage"}},{"name":"upcoming_only","in":"query","required":false,"schema":{"type":"boolean","default":false,"title":"Upcoming Only"}},{"name":"sport","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sport"}},{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/matches/{match_id}/picks":{"get":{"summary":"Get Match Picks","operationId":"get_match_picks_matches__match_id__picks_get","parameters":[{"name":"match_id","in":"path","required":true,"schema":{"type":"string","title":"Match Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/recommendations":{"get":{"summary":"Get Recommendations","operationId":"get_recommendations_recommendations_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":50,"minimum":1,"default":10,"title":"Limit"}},{"name":"sport","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sport"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/picks/recent":{"get":{"summary":"List Recent Picks","description":"Recent picks across all bet types, optionally filtered by sport or platform.","operationId":"list_recent_picks_picks_recent_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}},{"name":"sport","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sport"}},{"name":"platform","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Platform"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/picks/props":{"get":{"summary":"List Prop Picks","description":"Recent non-moneyline picks (draw, O/U, BTTS, props).","operationId":"list_prop_picks_picks_props_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}},{"name":"bet_type","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Bet Type"}},{"name":"sport","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Sport"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trading/calibration":{"get":{"summary":"Trading Calibration","description":"Model calibration summary: Brier score, hit rate by bucket, ROI.","operationId":"trading_calibration_trading_calibration_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/trading/paper":{"get":{"summary":"Trading Paper","description":"Virtual paper-trading bankroll summary (consensus-vs-self).","operationId":"trading_paper_trading_paper_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/trading/autobet":{"get":{"summary":"Trading Autobet","description":"Polymarket autobet performance + recent bets (consensus-vs-market).","operationId":"trading_autobet_trading_autobet_get","parameters":[{"name":"limit","in":"query","required":false,"schema":{"type":"integer","maximum":200,"minimum":1,"default":50,"title":"Limit"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/trading/autobet/run":{"post":{"summary":"Trading Autobet Run","description":"Manually trigger one autobet scan (respects paper/live mode + risk gates).","operationId":"trading_autobet_run_trading_autobet_run_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/stats/overview":{"get":{"summary":"Stats Overview","operationId":"stats_overview_stats_overview_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/stats/platforms":{"get":{"summary":"Stats Platforms","description":"Influencer and pick counts broken down by platform and sport.","operationId":"stats_platforms_stats_platforms_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/seed":{"post":{"summary":"Seed Influencers","description":"Populate the influencer list with curated accounts. Run once.","operationId":"seed_influencers_seed_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/sync":{"post":{"summary":"Manual Sync","description":"Trigger a full scrape + WC data sync cycle immediately.","operationId":"manual_sync_sync_post","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"},"input":{"title":"Input"},"ctx":{"type":"object","title":"Context"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}}}}